core/pyproject.toml
[project]
name = "job-search-api"
version = "0.1.0"
description = "A job search API using FastAPI and LangChain"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "langchain==0.0.350",
    "google-search-results==2.4.2",
    "python-dotenv==1.0.0",
    "pydantic==2.5.2",
    "pydantic-settings==2.1.0",
    "beautifulsoup4==4.12.2",
    "requests==2.31.0",
    "openai==1.12.0",
    "aiofiles==23.2.1",
    "python-multipart==0.0.6",
    "python-magic==0.4.27",
    "aiofiles.os==0.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core"]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I"]